I'm done. I've updated a lot. Removed all warnings. Created a release build. Removed the shop from level 1. Added debug flags everywhere. Made map level update if you switch cheats on or off. Disabled the ability to activate debug mode in release builds. Removed flying fox from release build. Polished up the shop 1 so we can show it. Updated the splash screens for show as well.

+++ MansionMashers/MansionMashers/SourceFiles/Enemy.c
@@ -121,7 +121,7 @@ Enemy* CreateEnemy(int enemyType, int collisionGroup, int objID, float xPos, flo
 		CurrentEnemy->EnemyRigidBody.onGround	= FALSE;
 		CurrentEnemy->dropDown					= FALSE;
 
-		InitializeEnemyStats(CurrentEnemy, 50, 250 + 10 * (rand() % 10), 15, 0, 10, 10, 10);
+		InitializeEnemyStats(CurrentEnemy, 50, (float)(250 + 10 * (rand() % 10)), 15.0f, 0, 10, 10, 10);
 
 		CurrentEnemy->EnemyParticleSystem = CreateFoxParticleSystem("TextureFiles/Particle.png", CurrentEnemy->Position.x / GetLoadRatio(), CurrentEnemy->Position.y / GetLoadRatio(), CurrentEnemy->EnemySprite->ZIndex + 5, 0, 5, 0.0f, 0, 360, 1.0f, -5.0f, 25, 24, 20, 2.0f, 0.5f);
